# docs/openapi/enrollment.json
{
  "components": {
    "schemas": {
      "Body_enroll_user_enrollment_enroll_post": {
        "properties": {
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "User's age (optional)",
            "title": "Age"
          },
          "photos": {
            "description": "Exactly 5 photos (JPG/PNG, max 5MB each)",
            "items": {
              "format": "binary",
              "type": "string"
            },
            "title": "Photos",
            "type": "array"
          },
          "relation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Relationship (e.g., 'Father', 'Sister') (optional)",
            "title": "Relation"
          },
          "user_name": {
            "description": "Name of the user",
            "title": "User Name",
            "type": "string"
          },
          "voice_samples": {
            "description": "Exactly 5 voice samples (WAV/MP3, max 10MB each)",
            "items": {
              "format": "binary",
              "type": "string"
            },
            "title": "Voice Samples",
            "type": "array"
          }
        },
        "required": [
          "user_name",
          "photos",
          "voice_samples"
        ],
        "title": "Body_enroll_user_enrollment_enroll_post",
        "type": "object"
      },
...
      "EnrollmentResponse": {
        "properties": {
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Token"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "expires_in": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "token_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Type"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "EnrollmentResponse",
        "type": "object"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "status_code": {
            "type": "integer"
          }
        },
        "required": [
          "status_code",
          "code",
          "message",
          "request_id"
        ],
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\n    ## Enrollment Service for Project NEXI - Week 3 Enhanced\n    \n    This service handles the complete user enrollment workflow with:\n    -  Photo capture and validation\n    -  Voice sample capture and validation\n    -  AI-powered face embedding extraction (via Vision Service)\n    -  AI-powered voice embedding extraction (via Audio Service)\n    -  User registration with Central Server\n    -  **Secure encrypted storage for enrollment data** (NEW)\n    -  **Storage management and statistics** (NEW)\n    -  **Data retrieval and deletion capabilities** (NEW)\n    \n    ### Service Architecture\n    - **Port:** 8005\n    - **Dependencies:** \n      - Vision Service (8001)\n      - Audio Service (8002)\n      - Central Server (8000)\n    \n    ### Week 3 Features\n    [OK] Secure encrypted storage  \n    [OK] Integration with Central Server user profiles  \n    [OK] Storage statistics and management  \n    [OK] Data retrieval and deletion APIs  \n    [OK] Enhanced progress tracking  \n    \n    ### Security\n    - All enrollment data is encrypted at rest\n    - Encryption keys are securely generated and stored\n    - Audit logging for all operations\n    ",
    "title": "NEXI Enrollment Service",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/enrollment/enroll": {
      "post": {
        "description": "Enroll a new user with 5 photos and 5 voice samples",
        "operationId": "enroll_user_enrollment_enroll_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_enroll_user_enrollment_enroll_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response"
          }
        },
        "summary": "Enroll User",
        "tags": [
          "Enrollment"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local service",
      "url": "http://localhost:8005"
    }
  ]
}